Authentication and Authorization using PAssport-JWt

+++ controllers/api/v1/posts_api.js
@@ -1,6 +1,6 @@
-const Post= require('../../../models/post');
-const Comment= require('../../../models/comment');
-module.exports.index= async function(req,res){
+const Post = require('../../../models/post');
+const Comment = require('../../../models/comment');
+module.exports.index = async function(req, res){
 
 
     let posts = await Post.find({})
@@ -13,33 +13,38 @@ module.exports.index= async function(req,res){
             }
         });
 
-
-    return res.json(200,{
-        message: "Lists of Posts",
+    return res.json(200, {
+        message: "List of posts",
         posts: posts
     })
 }
+
 
 module.exports.destroy = async function(req, res){
 
     try{
         let post = await Post.findById(req.params.id);
 
-        // if (post.user == req.user.id){
+        if (post.user == req.user.id){
             post.remove();
 
             await Comment.deleteMany({post: req.params.id});
 
-            req.flash('success', 'Post and associated comments deleted!');
 
-            return res.json(200,{
-                message:"Posts and associated comments deleted successfully"
+    
+            return res.json(200, {
+                message: "Post and associated comments deleted successfully!"
             });
+        }else{
+            return res.json(401, {
+                message: "You cannot delete this post!"
+            });
+        }
 
     }catch(err){
-        console.log('*******', err);
-        return res.json(500,{
-            message: "internal server Error"
+        console.log('********', err);
+        return res.json(500, {
+            message: "Internal Server Error"
         });
     }
